Create README.md
This is an AI based business chat-bot. particularly specialized in Medical information. It gives the disease name when all symptoms are given using AI methodologies.It also describes the disease and states the preventive measure need to be taken. A virtual environment maybe needed to be setup using visual code to run this as it runs in the flask environment.